Refactor comparison ignoring genotypes using vcfeval built-in feature

+++ starfish.py
@@ -50,57 +50,42 @@ def count_records(vcf_fname):
     vcf = ps.VariantFile(vcf_fname, 'r')
     return sum(1 for rec in vcf)
 
-def has_genotypes(vcf_name):
-    vcf = ps.VariantFile(vcf_name)
-    return vcf.header is not None and len(vcf.header.samples) > 0
-
-def drop_genotypes(vcf_fname):
-    tmp_vcf_fname = vcf_fname.replace('.vcf', '.drop_gt_tmp.vcf')
-    bcftools_cmd = ['bcftools', 'view', '-G', '-Oz', '-o', tmp_vcf_fname, vcf_fname]
-    call(bcftools_cmd)
-    os.rename(tmp_vcf_fname, vcf_fname)
-    if vcf_index_exists(vcf_fname):
-        index_vcf(vcf_fname)
-
-def reset_genotypes(in_vcf_name, out_vcf_name, sample, gt='0/1'):
-    bcftools_cmd = ['bcftools', 'view', '-G', in_vcf_name]
-    add_gt_awk_expression = 'BEGIN {FS="\t"; OFS=FS;} {if (NF < 5) print; else if ($1=="#CHROM") print $0, "FORMAT", "' + sample + '"; else print $0, "GT", "' + gt + '";}'
-    add_gt_awk_cmd = ['awk', add_gt_awk_expression]
-    add_gt_hdr_line_awk_expression = '!found && /#CHROM/{print "##FORMAT=<ID=GT,Number=1,Type=String,Description=\"Genotype\">"; found=1}1'
-    add_gt_hdr_line_awk_cmd = ['awk', add_gt_hdr_line_awk_expression]
-    bgzip_cmd = ['bgzip']
-    out_vcf = open(out_vcf_name, 'w')
-    bcftools = sp.Popen(bcftools_cmd, stdout=sp.PIPE)
-    add_gt_awk = sp.Popen(add_gt_awk_cmd, stdin=bcftools.stdout, stdout=sp.PIPE)
-    add_gt_hdr_awk = sp.Popen(add_gt_hdr_line_awk_cmd, stdin=add_gt_awk.stdout, stdout=sp.PIPE)
-    bgzip = sp.Popen(bgzip_cmd, stdin=add_gt_hdr_awk.stdout, stdout=out_vcf)
-    bcftools.stdout.close()
-    output = bgzip.communicate()[0]
-    return out_vcf_name
-
 def run_rtg(rtg, ref_sdf, lhs_vcf, rhs_vcf, out_dir,
-            bed_regions=None, all_records=False, squash_ploidy=False,
-            sample=None, threads=None, score_field='QUAL'):
+            bed_regions=None, all_records=False,
+            ref_overlap=False, squash_ploidy=False,
+            ignore_genotypes=False, sample=None,
+            score_field='QUAL', threads=None):
     cmd = [rtg, 'vcfeval', '-t', ref_sdf, '-b', lhs_vcf, '-c', rhs_vcf, '-o', out_dir]
     if bed_regions is not None:
         cmd += ['--bed-regions', bed_regions]
     if all_records:
         cmd.append('--all-records')
+    if ref_overlap:
+        cmd.append('--ref-overlap')
     if squash_ploidy:
         cmd.append('--squash-ploidy')
-    if sample is not None:
+    if ignore_genotypes:
+        cmd += ['--sample', 'ALT']
+    elif sample is not None:
         cmd += ['--sample', sample]
     if threads is not None:
         cmd += ['--threads', str(threads)]
     cmd += ['--vcf-score-field', score_field]
     call(cmd)
 
-def intersect(lhs_label, lhs_vcf, rhs_label, rhs_vcf, args):
+def make_empty_vcf(vcf_fname, template_vcf_fname, index=True):
+    bcftools_cmd = ['bcftools', 'view', '-h', '-Oz', '-o', vcf_fname, template_vcf_fname]
+    call(bcftools_cmd)
+    if index:
+        index_vcf(vcf_fname)
+
+def rtg_intersect(lhs_label, lhs_vcf, rhs_label, rhs_vcf, args):
     tmp_dir = join(args.output, 'temp')
     run_rtg(args.rtg, args.sdf, lhs_vcf, rhs_vcf, tmp_dir,
             bed_regions=args.regions,
             all_records=args.all_records,
-            squash_ploidy=args.ignore_genotypes,
+            ref_overlap=args.ref_overlap,
+            squash_ploidy=args.sample == "ALT",
             sample=args.sample,
             threads=args.threads)
     lhs_and_rhs = join(args.output, lhs_label + '_and_' + rhs_label + '.vcf.gz')
@@ -111,14 +96,23 @@ def intersect(lhs_label, lhs_vcf, rhs_label, rhs_vcf, args):
     tp = join(tmp_dir, 'tp.vcf.gz')
     fn = join(tmp_dir, 'fn.vcf.gz')
     fp = join(tmp_dir, 'fp.vcf.gz')
-    shutil.move(tp_baseline, lhs_and_rhs)
-    shutil.move(tp_baseline + '.tbi', lhs_and_rhs + '.tbi')
-    shutil.move(tp, rhs_and_lhs)
-    shutil.move(tp + '.tbi', rhs_and_lhs + '.tbi')
-    shutil.move(fn, lhs_not_rhs)
-    shutil.move(fn + '.tbi', lhs_not_rhs + '.tbi')
-    shutil.move(fp, rhs_not_lhs)
-    shutil.move(fp + '.tbi', rhs_not_lhs + '.tbi')
+    if exists(tp_baseline):
+        shutil.move(tp_baseline, lhs_and_rhs)
+        shutil.move(tp_baseline + '.tbi', lhs_and_rhs + '.tbi')
+        shutil.move(tp, rhs_and_lhs)
+        shutil.move(tp + '.tbi', rhs_and_lhs + '.tbi')
+        shutil.move(fn, lhs_not_rhs)
+        shutil.move(fn + '.tbi', lhs_not_rhs + '.tbi')
+        shutil.move(fp, rhs_not_lhs)
+        shutil.move(fp + '.tbi', rhs_not_lhs + '.tbi')
+    else:
+        # then there were no sequences in common (or one/both of the input are empty)
+        make_empty_vcf(lhs_and_rhs, lhs_vcf)
+        make_empty_vcf(rhs_and_lhs, rhs_vcf)
+        shutil.copyfile(lhs_vcf, lhs_not_rhs)
+        shutil.copyfile(lhs_vcf + '.tbi', lhs_not_rhs + '.tbi')
+        shutil.copyfile(rhs_vcf, rhs_not_lhs)
+        shutil.copyfile(rhs_vcf + '.tbi', rhs_not_lhs + '.tbi')
     shutil.rmtree(tmp_dir)
     return lhs_and_rhs, rhs_and_lhs, lhs_not_rhs, rhs_not_lhs
 
@@ -145,9 +139,12 @@ def naive_intersect(head_vcf, tail_vcfs, tail_mask, out):
     else:
         naive_common([head_vcf] + positive_tails, out)
 
-def concat(vcfs, out):
+def concat(vcfs, out, remove_duplicates=True):
     assert len(vcfs) > 1
-    cmd = ['bcftools', 'concat', '-a', '-Oz', '-o', out] + vcfs
+    cmd = ['bcftools', 'concat', '-a', '-Oz', '-o', out]
+    if remove_duplicates:
+        cmd.append('-D')
+    cmd += vcfs
     call(cmd)
 
 def plot_ven(labels, names):
@@ -176,17 +173,8 @@ def main(args):
     
     if not exists(args.output):
         os.makedirs(args.output)
-    
-    vcfs, temp_vcfs = [], []
-    for vcf in args.variants:
-        if args.ignore_genotypes:
-            tmp_vcf = join(args.output, basename(vcf).replace('.vcf', '.rtg.vcf'))
-            reset_genotypes(vcf, tmp_vcf, 'SAMPLE' if args.sample is None else args.sample)
-            index_vcf(tmp_vcf)
-            temp_vcfs.append(tmp_vcf)
-            vcf = tmp_vcf
-        vcfs.append(vcf)
-    
+        
+    vcfs = args.variants
     labels = string.ascii_uppercase[:len(vcfs)]
     
     if len(vcfs) > len(labels):
@@ -200,11 +188,7 @@ def main(args):
     for (lhs_label, lhs_vcf), (rhs_label, rhs_vcf) in itertools.combinations(labelled_vcfs, 2):
         if lhs_label not in intersections:
             intersections[lhs_label] = {}
-        intersection = intersect(lhs_label, lhs_vcf, rhs_label, rhs_vcf, args)
-        if args.ignore_genotypes:
-            for vcf_fname in intersection:
-                drop_genotypes(vcf_fname)
-        intersections[lhs_label][rhs_label] = intersection
+        intersections[lhs_label][rhs_label] = rtg_intersect(lhs_label, lhs_vcf, rhs_label, rhs_vcf, args)
     
     for i in range(len(vcfs)):
         isecs = [intersections[labels[i]][rhs_label][0] for rhs_label in labels[i + 1:]]
@@ -220,10 +204,8 @@ def main(args):
                         isecs.append(intersections[head][labels[i]][1])
                         tail_mask.append(0)
                 naive_intersect(vcfs[i], isecs, tail_mask, out_vcf)
-                if args.ignore_genotypes:
-                    drop_genotypes(out_vcf)
                 index_vcf(out_vcf)
-    
+        
     # Find unique records
     for i, label in enumerate(labels):
         isecs = []
@@ -294,10 +276,6 @@ def main(args):
                 plt.savefig(args.vennout, format='pdf', )
         else:
             print("Venn plots only supported for up to 6 VCFs")
-    
-    # Cleanup
-    for vcf in temp_vcfs:
-        remove_vcf(vcf)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
@@ -322,10 +300,6 @@ if __name__ == '__main__':
                         type=str,
                         required=False,
                         help='regions in BED format to perform intersection')
-    parser.add_argument('--ignore_genotypes',
-                        default=False,
-                        action='store_true',
-                        help='Ignore genotypes during intersection; use variants only')
     parser.add_argument('--all_records',
                         default=False,
                         action='store_true',
@@ -333,7 +307,7 @@ if __name__ == '__main__':
     parser.add_argument('--sample',
                         type=str,
                         required=False,
-                        help='Sample to compare, only required if multiple samples in VCFs')
+                        help='Sample to compare (if multiple samples in VCFs) or ALT to ignore genotypes')
     parser.add_argument('--names',
                         nargs='+',
                         type=str,
@@ -347,5 +321,9 @@ if __name__ == '__main__':
                         type=int,
                         required=False,
                         help='Maximum number of threads to use (default is all cores)')
+    parser.add_argument('--ref_overlap',
+                        default=False,
+                        action='store_true',
+                        help='Call RTG vcfeval with "ref-overlap" option')
     parsed, unparsed = parser.parse_known_args()
     main(parsed)
